+++ impractical/simple-nat/shared/src/main/scala/net/scalax/simple/nat/group01/number_04/Number.scala
@@ -1,13 +1,17 @@
-package net.scalax.simple.nat.number19
+package net.scalax.simple.nat.group01.number_04
 
 import net.scalax.simple.ghdmzsk.ghdmzsk
 
 object Number {
 
-  val number1S: ghdmzsk = ghdmzsk(num1Tail => ghdmzsk(num2 => AA(() => num1Tail()(num2))))
-  val number1T: ghdmzsk = ghdmzsk(num1Tail => ghdmzsk(num2 => num2()(num1Tail)))
-  val number3S: ghdmzsk = ghdmzsk(num3Tail => ghdmzsk(num2 => num2()(num3Tail)))
-  val number3T: ghdmzsk = ghdmzsk(num3Tail => ghdmzsk(num1 => BB(() => num1()(num3Tail))))
+  val number1S: ghdmzsk =
+    ghdmzsk(num1Tail => ghdmzsk(num2 => ghdmzsk(num3 => ghdmzsk(num4 => ghdmzsk(num5 => AA(() => num1Tail()(num2)(num3)(num4)(num5)))))))
+  val number1T: ghdmzsk =
+    ghdmzsk(num1Tail => ghdmzsk(num2 => ghdmzsk(num3 => ghdmzsk(num4 => ghdmzsk(num5 => num2()(num3)(num4)(num5)(num1Tail))))))
+  val number3S: ghdmzsk =
+    ghdmzsk(num2Tail => ghdmzsk(num3 => ghdmzsk(num4 => ghdmzsk(num5 => ghdmzsk(num1 => num1()(num2Tail)(num3)(num4)(num5))))))
+  val number3T: ghdmzsk =
+    ghdmzsk(num5Tail => ghdmzsk(num1 => ghdmzsk(num2 => ghdmzsk(num3 => ghdmzsk(num4 => BB(() => num1()(num2)(num3)(num4)(num5Tail)))))))
 
   case class AA(cc: () => ghdmzsk) extends ghdmzsk {
     override def apply(t: () => ghdmzsk): ghdmzsk = throw new Exception
