+++ impractical/simple-nat/shared/src/main/scala/net/scalax/simple/nat/group01/number_05/RunMain.scala
@@ -1,15 +1,16 @@
-package net.scalax.simple.nat.number17
+package net.scalax.simple.nat.group01.number_05
 
 import net.scalax.simple.ghdmzsk.ghdmzsk
 
 import scala.annotation.tailrec
 
-object `MeMeMe我我我` extends App {
+object `喵_Number_05` extends App {
 
   def genNumber(c: Int): (ghdmzsk, ghdmzsk, ghdmzsk) = {
-    lazy val (n1Pos, _): (ghdmzsk, ghdmzsk) = Number.genNumberImpl(positive = Number.number1S, zero = Number.number2T, count = c - 1)
-    lazy val (n2Pos, _): (ghdmzsk, ghdmzsk) = Number.genNumberImpl(positive = Number.number2S, zero = Number.number2T, count = c)
-    lazy val (n3Pos, _): (ghdmzsk, ghdmzsk) = Number.genNumberImpl(positive = Number.number2S, zero = Number.number2U, count = c)
+    lazy val (n1Pos, _): (ghdmzsk, ghdmzsk) = Number.genNumberImpl(positive = Number.number1S, zero = Number.number1T, count = c - 1)
+    lazy val (n2Pos, _): (ghdmzsk, ghdmzsk) = Number.genNumberImpl(positive = Number.number3S, zero = Number.number1T, count = c)
+    lazy val (n3Pos, _): (ghdmzsk, ghdmzsk) = Number.genNumberImpl(positive = Number.number3S, zero = Number.number3T, count = c)
+
     (n1Pos, n2Pos, n3Pos)
   }
 
@@ -19,7 +20,7 @@ object `MeMeMe我我我` extends App {
     var length2: Long   = 0
 
     val (n1, n2, n3) = genNumber(except)
-    val gen          = () => n1(() => n2)(() => n2)(() => n3)
+    val gen          = () => n1(() => n2)(() => n2)(() => n2)(() => n2)(() => n3)
 
     @tailrec
     def countImpl(num: () => ghdmzsk, needConitie: Long): Unit = {
@@ -31,7 +32,9 @@ object `MeMeMe我我我` extends App {
           length2 = length2 + 1
           tail
       }
+
       def r = BigDecimal(length1) / BigDecimal(length2)
+
       if (needConitie > 0)
         countImpl(newTail, needConitie = needConitie - 1)
       else if (length2 == 0)
@@ -45,13 +48,13 @@ object `MeMeMe我我我` extends App {
     (length1, length2)
   }
 
-  for (i <- 1 to 10) {
-    val i1: Long       = i * i * i * i
+  for (i <- 1 to 6) {
+    val i1: Long       = i * i * i * i * i * i
     val (g1, g2)       = count(except = i, i1 = i1)
     val i2: BigDecimal = BigDecimal(i1 - 1)
     val r: BigDecimal  = BigDecimal(g1) / BigDecimal(g2)
     println(g1, g2)
-    println(i, s"$i ^ 4 - 1", r)
+    println(i, s"$i ^ 6 - 1", r)
     assert((i2 - r).abs < BigDecimal("0.001"))
   }
